requirements.txt
absl-py==2.1.0
addict==2.4.0
aliyun-python-sdk-core==2.16.0
aliyun-python-sdk-kms==2.16.5
annotated-types==0.7.0
av==13.1.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
contourpy==1.3.0
crcmod==1.7
cryptography==43.0.1
cycler==0.12.1
deepspeed==0.16.0
einops==0.8.0
facexlib==0.3.0
filelock==3.14.0
filterpy==1.4.5
fonttools==4.54.1
fsspec==2024.9.0
fvcore==0.1.5.post20221221
grpcio==1.67.0
hjson==3.1.0
huggingface-hub==0.25.2
idna==3.10
importlib_metadata==8.5.0
iopath==0.1.10
Jinja2==3.1.4
jmespath==0.10.0
kiwisolver==1.4.7
llvmlite==0.43.0
lmdb==1.5.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.1
matplotlib==3.9.2
mdurl==0.1.2
model-index==0.1.11
mpmath==1.3.0
msgpack==1.1.0
natten==0.14.6+torch1130cu117
networkx==3.4.2
ninja==1.11.1.2
numba==0.60.0
numpy==1.23.5
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-ml-py==12.560.30
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
opencv-python==4.10.0.84
opendatalab==0.0.10
openmim==0.3.9
openxlab==0.1.2
ordered-set==4.1.0
oss2==2.17.0
packaging==24.1
pandas==2.2.3
pillow==11.0.0
platformdirs==4.3.6
portalocker==3.0.0
protobuf==5.28.2
psutil==6.1.0
py-cpuinfo==9.0.0
pycparser==2.22
pycryptodome==3.21.0
pydantic==2.10.2
pydantic_core==2.27.1
Pygments==2.18.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
pytorch-msssim==1.0.0
pytz==2023.4
PyYAML==6.0.2
requests==2.28.2
rich==13.4.2
safetensors==0.4.5
scipy==1.14.1
six==1.16.0
sympy==1.13.1
tabulate==0.9.0
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorboardX==2.6.2.2
termcolor==2.5.0
timm==1.0.11
tomli==2.0.2
torch==1.13.0+cu117
torch-flops==0.3.6
torchaudio==0.13.0+cu117
torchvision==0.14.0+cu117
tqdm==4.65.2
triton==3.1.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==1.26.20
Werkzeug==3.0.4
yacs==0.1.8
yapf==0.40.2
zipp==3.20.2
